separating processed expression and factor, force empty one-tensor
in order to be able to directly used the processed expression
elsewhere (e.g. tal-sh, exacorr, exatn, for which the factor
is given as a scaling factor, but not in the string representing
the tensor operation) we do not add the factor
multiplying the expressions (original, or recalculated after
splitting groups)  to the saved processed expression.

however, since we keep the operators, factor and the processed
expression, we can form the final expression whenever desired.

in addition to that, again aiming at producing expressions for
tal-sh, exacorr, exatn, we set as default behavior for rank1 tensors
the removal of the middle bar, by defining as default an override argument
(empty_unit_tensor=True) to the tensor.set_tensor_representation()
method.