WIP saving work or classifying tensors
add code to check whether a given tensor is input (only appears on rhs),
output (only appears on lhs), or "intermediate" (appears both on rhs and lhs),
and sort these into lists.

right not this information seems important to me when writing code,
since ideally we want to create variable declarations, and whether they
are arguments to the subroutine implementing the sequence of tensor
operations, or are internal to the routine.

currently the sorting not working since i'm checking on tensor objects, but i
should be checking on tensor names as well; probably not only the name but
all characteristics of the tensor (rank, arguments etc).

this is because in the way the sial input file is being parse, different
tensor objects will generally reference a particular tensor name.