saving work on supporting complex conjugation in code generation
split the name from the arguments in parenthesis to allow for eventually
appending the complex conjugation sign on the tensor name.

still not certain on where is the best place to operate this change, but should
be when writing the code: a given tensor in the contraction has to have is printed
representation changed if index swaping and/or complex conjugation is asked for,
to conform with the available/allowed tensor classes. or perhaps not...