introducing option to get groups as collections of indexes or many-body indexes
the identification of repeating groups in binary tensor contractions
relies on comparing the many-body tensor indexes, and not their indexes
in the string defining the tensor, which was introduced in the previous
commit.

now the default behavior is to get this information from a tensor instance
in terms of the many-body quantities.