splitting parsing from validation
splitting the parsing of sial input lines from their validation, with
validation here meaning whether a given sial input line conforms to
the constraints we wish to apply: for instance, getting to spatial orbitals
to spin orbitals/spinors, and subsequently, restrictions on the class (such
as only having vo tensors, and thus replacing any ov tensors by the complex
conjugate of the vo tensor).

with this splitting we retain the complete information read from the sial
input, and use the validated copy, which may have been modified by the constraints.