reusing previously created instances of tensors
added to the the SIAL processing, as well as to the contraction
and tensor processing, a verification on whether a particular
label, known at a given point of input processing, is associated with
an existing instance of a tensor.

if such an instance exists, associate it with the instruction being processed,
otherwise create and process a new instance..

it may be that in some cases, a definition of indexes/groups changes from
the CREATE command to the CONTRACTION one. in this case, the content of the
tensor is reset and it is reinitialize to the form found in the contraction.

In some cases, that may change index names or ordering. This does not appear
to cause isses but if that is found to be the case, one should revert to the
case of creating new instances for each operation.