modified layout of events
now when invoking get_event for different instructions, data is kept
in the following format

for tensors:
{ LHS tensor name : [instruction line number, action, tensor instance] }

for contractions (C = A*B):
{ LHS tensor name : [instruction line number, action, contraction instance, { C tensor name : C tensor instance}, { A tensor name : A tensor instance}, { B tensor name : B tensor instance}] }

where B is optional; for the moment I keep the dictionaries with tensor names and instances in the list, since it looks more convenient to see which tensors are part of the contraction, though arguably that's redundant as all information is par of the contraction instance. I'll see whether to keep that or not.

one idea now is to construct the lifetimes / dependency trees from a list of all events