adding class to handle infromation on diagrams
now the comments on the sial file indicating to which
diagrams a particular contraction is associated to are
collected and parsed.

however, no use is made of such information yet.

at the sial parsing level, as we know the lines with diagram
informaton and the contraction follow each other, the idea
is to then cross-link the "diagram info" objects and the contraction
objects so that from one one can retrieve information on the other.

the goal of keeping diagram information is to see whether that
can be, at the code generation stage, used to eliminate, or transform
contractions connected to particular diagrams.