insert regex marker for EOL in contract/add  re object construction
this commit fixes an issue found by jan brandejs when parsing
an unary contraction with a factor, e.g.

A8(f1al1a|n1am1a)+=I1(l1af1a|n1am1a)*4.

in which the code was incorrectly considering 4. to be a tensor,
as it identified the line as a binary contraction instead of a
unary operation (scaling by a factor).

now, by adding a regex character for end of line, there seems to
be no confusion between the four possible cases we can encounter,
that is

- binary contraction
- binary contraction with a scaling factor
- unary operation (e.g. tensor addition)
- unary operation with a scaling factor